+++ src/Morphic/SceneMorph.cpp
@@ -56,7 +56,7 @@ void SceneMorph::handleMouseMotionEvent(const MouseMotionEventPtr &event)
 {
     if(event->hasLeftButtonDown())
     {
-        cameraAngles += Vector3(-event->delta.y, -event->delta.x, 0)*(M_PI/180.0);
+        cameraAngles += Vector3(-event->delta.y, -event->delta.x, 0)*float(M_PI/180.0);
         cameraNode->transform.rotation = computeCameraOrientation();
         //printf("Scene morph left  %f,%f\n", event->delta.x, event->delta.y);
     }
@@ -65,7 +65,7 @@ void SceneMorph::handleMouseMotionEvent(const MouseMotionEventPtr &event)
     {
         cameraNode->transform.translation +=
             cameraNode->transform.rotation.rotateVector(
-                Vector3(event->delta.x, -event->delta.y, 0).normalized()*0.05
+                Vector3(event->delta.x, -event->delta.y, 0).normalized()*0.05f
             );
     }
 
@@ -76,7 +76,7 @@ void SceneMorph::handleMouseWheelEvent(const MouseWheelEventPtr &event)
     //printf("Wheel %f %f\n", event->position.x, event->position.y);
     cameraNode->transform.translation += 
         cameraNode->transform.rotation.rotateVector(
-            Vector3(event->scrollAmount.x, 0, -event->scrollAmount.y)*0.1
+            Vector3(event->scrollAmount.x, 0, -event->scrollAmount.y)*0.1f
         );
 }
